+++ bindings/react/demo/index.js
@@ -35,6 +35,8 @@ import RangeExample from './examples/Range';
 import RowColumnExample from './examples/RowColumn';
 import BackButtonExample from './examples/BackButton';
 import BottomToolbarExample from './examples/BottomToolbar';
+import RouterNavigatorExample from './examples/RouterNavigator';
+import ReduxNavigatorExample from './examples/ReduxNavigator';
 
 class Examples extends React.Component {
   constructor(props) {
@@ -132,6 +134,14 @@ class Examples extends React.Component {
       {
         title: 'Input',
         component: InputExample
+      },
+      {
+        title: 'Stateless Navigator',
+        component: RouterNavigatorExample
+      },
+      {
+        title: 'Redux Navigator',
+        component: ReduxNavigatorExample
       }
     ];
   }