+++ core/src/elements/ons-fab.spec.js
@@ -27,6 +27,15 @@ describe('OnsFabElement', () => {
     expect(fab.classList.contains('fab--fuga')).to.be.true;
   });
 
+  onlyChrome(describe)('"class" attribute', () => {
+    it('should contain "fab" token automatically', () => {
+      expect(fab.classList.contains('fab')).to.be.true;
+      fab.setAttribute('class', 'foo');
+      expect(fab.classList.contains('fab')).to.be.true;
+      expect(fab.classList.contains('foo')).to.be.true;
+    });
+  });
+
   describe('#_show()', () => {
     it('calls show()', () => {
       const spy = chai.spy.on(fab, 'show');